minor clarification and simplification
"with" could be misread as "using", and the language in Query
Constraints ("scores made by any player in a particular list") is more
natural. With this replacement I have replaced "made" with "achieved" to
avoid potentially meaning "created by". The example below it clarifies
that a player can be specified using "playerName"